keep as uint8

+++ experiments/Trees/prepare_field_data.py
@@ -24,7 +24,6 @@ from time import sleep
 def resize(img, height, width):
     # resize image
     dim = (width, height)    
-    img = img.astype("float32")
     resized = cv2.resize(img, dim, interpolation = cv2.INTER_NEAREST)
     
     return resized
@@ -254,8 +253,8 @@ def create_records(HSI_crops, RGB_crops, labels, sites, heights, elevations, box
         chunk_elevations = elevations[i:i + chunk_size]
         chunk_heights = heights[i:i + chunk_size]
             
-        resized_RGB_crops = [resize(x, RGB_size, RGB_size).astype("float32") for x in chunk_RGB_crops]
-        resized_HSI_crops = [resize(x, HSI_size, HSI_size).astype("float32") for x in chunk_HSI_crops]
+        resized_RGB_crops = [resize(x, RGB_size, RGB_size) for x in chunk_RGB_crops]
+        resized_HSI_crops = [resize(x, HSI_size, HSI_size) for x in chunk_HSI_crops]
         
         filename = "{}/field_data_{}.tfrecord".format(savedir, counter)
         write_tfrecord(filename=filename,
